requirements.txt
black==22.3.0
click==8.1.2
clickhouse-driver==0.2.3
colorama==0.4.4
et-xmlfile==1.1.0
greenlet==1.1.2
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
pathspec==0.9.0
platformdirs==2.5.2
protobuf==3.20.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.35
tomli==2.0.1
tzdata==2022.1
tzlocal==4.2
